All features finally "work" on Android
The app now completely works on android with a trick. When taking the picture the capture button must be held until the picture is captured and saved. For some reason it will not wait for the image to be saved before trying to open it. If you only tap the capture button it tries to open an image that doesn't exist because it hasn't finishied saving yet.

A possible work arround could be to directly take the camera capture and store it as a local varriable. On my phone the image is captured as a bytearray object by jnius. This will need to be converted to an image array.

The camera preview is still rotated by 90 degrees. More digging is required to find out why.